+++ test/opf-var.jl
@@ -941,4 +941,56 @@ end
             @test isapprox(result["solution"]["branch"]["3"]["ta"], 5.0/180*pi; atol = 1e-1)
         end
     end
+
+
+    @testset "test dc polar opf" begin
+        @testset "3-bus case with fixed phase shift / tap" begin
+            file = "../test/data/matpower/case3_oltc_pst.m"
+            data = PowerModels.parse_file(file)
+            result = PowerModels.run_opf(data, DCPPowerModel, milp_solver)
+
+            @test result["termination_status"] == OPTIMAL
+            @test isapprox(result["objective"], 5782.0; atol = 1e0)
+        end
+
+        @testset "3-bus case with optimal phase shifting / tap changing" begin
+            file = "../test/data/matpower/case3_oltc_pst.m"
+            data = PowerModels.parse_file(file)
+            result = PowerModels._solve_opf_oltc_pst(data, DCPPowerModel, milp_solver)
+
+            @test result["termination_status"] == OPTIMAL
+            @test isapprox(result["objective"], 5639.0; atol = 1e0)
+
+            @test haskey(result["solution"]["branch"]["1"], "ta")
+
+            @test isapprox(result["solution"]["branch"]["1"]["tm"], 1.000; atol = 1e-3)
+            @test isapprox(result["solution"]["branch"]["1"]["ta"], 0.000; atol = 1e-3)
+            @test isapprox(result["solution"]["branch"]["2"]["tm"], 1.000; atol = 1e-3)
+            @test isapprox(result["solution"]["branch"]["2"]["ta"], 0.000; atol = 1e-3)
+            @test isapprox(result["solution"]["branch"]["3"]["tm"], 1.000; atol = 1e-3)
+            @test isapprox(result["solution"]["branch"]["3"]["ta"], 15.0/180*pi; atol = 1e-1)
+        end
+
+
+        @testset "3-bus case with optimal phase shifting / tap changing with equal lb/ub" begin
+            file = "../test/data/matpower/case3_oltc_pst.m"
+            data = PowerModels.parse_file(file)
+            for (i, branch) in data["branch"]
+                branch["ta_min"] = branch["shift"]
+                branch["ta_max"] = branch["shift"]
+            end
+            result = PowerModels._solve_opf_oltc_pst(data, DCPPowerModel, milp_solver)
+
+            @test result["termination_status"] == OPTIMAL
+            @test isapprox(result["objective"], 5698.1; atol = 1e0)
+
+            @test isapprox(result["solution"]["branch"]["1"]["tm"], 1.000; atol = 1e-3)
+            @test isapprox(result["solution"]["branch"]["1"]["ta"], 0.000; atol = 1e-3)
+            @test isapprox(result["solution"]["branch"]["2"]["tm"], 1.000; atol = 1e-3)
+            @test isapprox(result["solution"]["branch"]["2"]["ta"], 0.000; atol = 1e-3)
+            @test isapprox(result["solution"]["branch"]["3"]["tm"], 1.000; atol = 1e-3)
+            @test isapprox(result["solution"]["branch"]["3"]["ta"], 5.0/180*pi; atol = 1e-1)
+        end
+    end
+
 end